+++ DataBrowse/forms.py
@@ -1,8 +1,13 @@
 from django import forms
-from DataBrowse.models import DataFrameTIF
+from DataBrowse.models import DataObject
 
 
 class UploadDataFrameForm(forms.ModelForm):
     class Meta:
-        model = DataFrameTIF
-        fields = ('upload', 'data_frame_name', 'sequence_name')
+        model = DataObject
+        fields = ('upload',
+                  'data_frame_name',
+                  'sequence_name',
+                  'iteration_token',
+                  'object_type',
+                  'source')
