+++ DataBrowse/validators.py
@@ -2,6 +2,6 @@ def validate_file_extension(value):
     import os
     from django.core.exceptions import ValidationError
     ext = os.path.splitext(value.name)[1]  # [0] returns path+filename
-    valid_extensions = ['.tif', '.tiff']
+    valid_extensions = ['.tif', '.tiff', '.csv']
     if not ext.lower() in valid_extensions:
-        raise ValidationError(u'Unsupported file extension. Please upload a .tif or .tiff')
+        raise ValidationError(u'Unsupported file extension. Please upload a .tif / .tiff or .csv')
